requirements.txt
# Core dependencies
pypdf==4.0.1
groq==0.4.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.3.0
requests==2.31.0
huggingface-hub==0.19.4
supabase==2.3.5
python-multipart==0.0.6
python-dotenv==1.0.0

# FastAPI ecosystem (pinned to prevent resolver backtracking)
starlette==0.27.0
anyio==3.7.1
httpx>=0.24.1,<0.26
sniffio==1.3.1
typing-extensions==4.15.0
click==8.3.1
h11==0.16.0

# Uvicorn extras
httptools==0.7.1
pyyaml==6.0.3
# uvloop==0.22.1  # Not supported on Windows
watchfiles==1.1.1
websockets==15.0.1

# Pydantic
annotated-types==0.7.0
pydantic-core==2.6.3

# Requests & networking
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.2
certifi==2025.11.12

# HuggingFace
filelock==3.20.0
fsspec==2025.12.0
tqdm==4.67.1
packaging==25.0

# Supabase ecosystem (pinned versions that work together)
gotrue>=2.12.0,<3.0.0
postgrest>=0.15.0,<1.0.0
realtime>=1.0.0,<2.0.0
storage3>=0.7.0,<1.0.0
supafunc>=0.3.0,<1.0.0
pyjwt==2.10.1
httpcore>=0.17.0,<2.0.0
deprecation==2.1.0
strenum==0.4.15
python-dateutil==2.9.0.post0